+++ laktory/models/datasinks/filedatasink.py
@@ -81,6 +81,7 @@ class FileDataSink(BaseDataSink):
             self.merge_cdc_options.execute(
                 source=df,
                 target_path=self.path,
+                node=self,
             )
             return